// src/server/gatewayApi/generated/models/EventsIdPatchRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FHH VR - Backend API
 * This is the backend api for the FHHVR Project.
 *
 * The version of the OpenAPI document: 0.1.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EventsIdPatchRequest
 */
export interface EventsIdPatchRequest {
    /**
     * 
     * @type {string}
     * @memberof EventsIdPatchRequest
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof EventsIdPatchRequest
     */
    endTime?: string;
    /**
     * 
     * @type {string}
     * @memberof EventsIdPatchRequest
     */
    startTime?: string;
    /**
     * 
     * @type {string}
     * @memberof EventsIdPatchRequest
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof EventsIdPatchRequest
     */
    title?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof EventsIdPatchRequest
     */
    attendees?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof EventsIdPatchRequest
     */
    moderators: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof EventsIdPatchRequest
     */
    project?: string | null;
}

/**
 * Check if a given object implements the EventsIdPatchRequest interface.
 */
export function instanceOfEventsIdPatchRequest(value: object): value is EventsIdPatchRequest {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('moderators' in value) || value['moderators'] === undefined) return false;
    return true;
}

export function EventsIdPatchRequestFromJSON(json: any): EventsIdPatchRequest {
    return EventsIdPatchRequestFromJSONTyped(json, false);
}

export function EventsIdPatchRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): EventsIdPatchRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'endTime': json['endTime'] == null ? undefined : json['endTime'],
        'startTime': json['startTime'] == null ? undefined : json['startTime'],
        'status': json['status'] == null ? undefined : json['status'],
        'title': json['title'] == null ? undefined : json['title'],
        'attendees': json['attendees'] == null ? undefined : json['attendees'],
        'moderators': json['moderators'],
        'project': json['project'] == null ? undefined : json['project'],
    };
}

export function EventsIdPatchRequestToJSON(json: any): EventsIdPatchRequest {
    return EventsIdPatchRequestToJSONTyped(json, false);
}

export function EventsIdPatchRequestToJSONTyped(value?: EventsIdPatchRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'endTime': value['endTime'],
        'startTime': value['startTime'],
        'status': value['status'],
        'title': value['title'],
        'attendees': value['attendees'],
        'moderators': value['moderators'],
        'project': value['project'],
    };
}