// src/server/gatewayApi/generated/models/EventsPutRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FHH VR - Backend API
 * This is the backend api for the FHHVR Project.
 *
 * The version of the OpenAPI document: 0.1.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EventsPutRequest
 */
export interface EventsPutRequest {
    /**
     * 
     * @type {string}
     * @memberof EventsPutRequest
     */
    endTime: string;
    /**
     * 
     * @type {string}
     * @memberof EventsPutRequest
     */
    startTime: string;
    /**
     * 
     * @type {string}
     * @memberof EventsPutRequest
     */
    title: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof EventsPutRequest
     */
    attendees: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof EventsPutRequest
     */
    moderators: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof EventsPutRequest
     */
    project?: string;
}

/**
 * Check if a given object implements the EventsPutRequest interface.
 */
export function instanceOfEventsPutRequest(value: object): value is EventsPutRequest {
    if (!('endTime' in value) || value['endTime'] === undefined) return false;
    if (!('startTime' in value) || value['startTime'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('attendees' in value) || value['attendees'] === undefined) return false;
    if (!('moderators' in value) || value['moderators'] === undefined) return false;
    return true;
}

export function EventsPutRequestFromJSON(json: any): EventsPutRequest {
    return EventsPutRequestFromJSONTyped(json, false);
}

export function EventsPutRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): EventsPutRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'endTime': json['endTime'],
        'startTime': json['startTime'],
        'title': json['title'],
        'attendees': json['attendees'],
        'moderators': json['moderators'],
        'project': json['project'] == null ? undefined : json['project'],
    };
}

export function EventsPutRequestToJSON(json: any): EventsPutRequest {
    return EventsPutRequestToJSONTyped(json, false);
}

export function EventsPutRequestToJSONTyped(value?: EventsPutRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'endTime': value['endTime'],
        'startTime': value['startTime'],
        'title': value['title'],
        'attendees': value['attendees'],
        'moderators': value['moderators'],
        'project': value['project'],
    };
}